pyproject.toml
[tool.poetry]
name = "helm-charts-updater"
version = "0.1.0"
description = "A tool that will update helm charts in a given repository"
authors = ["Vadim Gedz <vadims@linux-tech.io>"]

[tool.poetry.dependencies]
python = "^3.10"
PyYAML = "^6.0"
semver = "^2.13.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
helm-charts-updater = "helm_charts_updater.main:main"
